[refactor] swagger 패키지 생성 후 관련 파일 위치 리팩토링 (#36)
