+++ src/main/java/com/sparta/vroomvroom/domain/address/controller/AddressController.java
@@ -5,8 +5,8 @@ import com.sparta.vroomvroom.domain.address.model.dto.response.AddressResponseDt
 import com.sparta.vroomvroom.domain.address.service.AddressService;
 import com.sparta.vroomvroom.domain.user.repository.UserRepository;
 import com.sparta.vroomvroom.global.conmon.BaseResponse;
-import com.sparta.vroomvroom.global.conmon.constants.SwaggerDescription;
-import com.sparta.vroomvroom.global.conmon.constants.SwaggerExamples;
+import com.sparta.vroomvroom.global.conmon.swagger.SwaggerDescription;
+import com.sparta.vroomvroom.global.conmon.swagger.SwaggerExamples;
 import com.sparta.vroomvroom.global.security.UserDetailsImpl;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
